Support dropping non-word headers
Most git tooling expects the first line to be a summary or a header of a
longer commit. In microblogs a heading is not always required, but
omitting it and writing the whole post in the first line(s) annoys
existing text editors with support for commit message highlighting/linting.

Instead we can now add a dummy header ("* * *", "...", "---") - and
provide a regex to drop such headers during rendering.